refactor: Restructuring the API
