+++ api/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "plexanibridge-mappings-api",
-	"version": "2.0.0",
+	"version": "2.0",
 	"private": false,
 	"scripts": {
 		"deploy": "wrangler deploy",
